+++ ai-services/assets/catalog/podman/values.yaml
@@ -1,6 +1,6 @@
 ui:
   port: ""
-  image: icr.io/ai-services-cicd/catalog-ui:v0.0.42
+  image: icr.io/ai-services-cicd/catalog-ui:v0.0.43
 
 backend:
   port: ""
