+++ ui/catalog/src/components/DeployFlow/Services/steps/StepZero.tsx
@@ -69,11 +69,8 @@ export const StepZero: React.FC<StepZeroProps> = ({
                   <div className={styles.serviceTileContent}>
                     <h3 className={styles.serviceTileName}>{service.name}</h3>
                     <div className={styles.certifiedBadge}>
-                      {/* <Checkmark size={16} /> */}
-                      <span className={styles.certifiedBadge}>
-                        <Badge size={16} className={styles.badgeIcon} />
-                        <span className={styles.badgeName}>IBM certified</span>
-                      </span>
+                      <Badge size={16} />
+                      <span>IBM certified</span>
                     </div>
                     <p className={styles.serviceTileDescription}>
                       {service.description}
