+++ ui/catalog/src/components/DeployFlow/Shared/_tokens.scss
@@ -1,0 +1,7 @@
+// DeployFlow token forwarder — no CSS output, forwards only.
+@forward "@carbon/react/scss/spacing";
+@forward "@carbon/react/scss/theme";
+@forward "@carbon/react/scss/type";
+@forward "@carbon/react/scss/breakpoint";
+@forward "@carbon/react/scss/motion";
+@forward "@carbon/colors";
